src/back/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "enteros-back"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.*",
    "uvicorn[standard]==0.32.*",
    "pydantic==2.9.*",
    "pydantic-settings==2.5.*",
    "sqlalchemy==2.0.*",
    "alembic==1.13.*",
    "psycopg[binary]==3.2.*",
    "pgvector==0.3.*",
    "openai==1.54.*",
    "sentence-transformers>=2.7.0",
    "pdfplumber==0.11.*",
    "pytesseract==0.3.*",
    "openpyxl==3.1.*",
    "pandas==2.2.*",
    "python-multipart==0.0.12",
    "python-jose[cryptography]==3.3.*",
    "passlib[bcrypt]==1.7.*",
    "httpx==0.27.*",
    "pyyaml==6.0.*",
    "torch>=2.1.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.*",
    "pytest-asyncio==0.23.*",
    "ruff==0.6.*",
    "mypy==1.11.*",
]

[tool.pytest.ini_options]
testpaths = ["tests", "app/tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
strict = true
